isomorphic SDFs: work around cargo linkage issues
cargo has problems properly prioritizing the shader rlib over the dylib,
creating linkage issues with duplicately-expressed dependencies when
the dependencies of the root crate intersect with dependencies of the
shader crate. by duplicating the shader crate and symlinking over the
code, we can avoid this issue by not specifying dylib generation in the
duplicate crate (`shaders-isomorphic`).